Revert "Move enums to correct package"
This reverts commit 25a43a5217ae4be9bf0822f039ef4b0c88b8b4e4.
