+++ models/src/main/java/com/vimeo/networking2/enums/SlackLanguagePreference.kt
@@ -1,6 +1,4 @@
-package com.vimeo.networking2.params
-
-import com.vimeo.networking2.enums.StringValue
+package com.vimeo.networking2.enums
 
 /**
  * Language preference of the Slack channel being notified.
